Only loop over arrays if there are conditions
If a type definition had an array of items that did not have any conditions,
this code would generate an empty loop, like so:

```c++
        for (const auto& item: masks) {
        }
```

This would cause compiler warnings in code that included this header due to
the unused variable and could potentially make serialization less efficient
if the compiler did not optimize that out.  This change will omit the
loop entirely if no conditions are generated.

Signed-off-by: P. J. Reed <phillipreed@hatchbed.com>